channels: reorg hook usage, change all brief usaqge
