+++ ui/src/heap/EditCurioForm.tsx
@@ -4,7 +4,6 @@ import { useDismissNavigate } from '@/logic/routing';
 import { EditCurioFormSchema } from '@/types/heap';
 import { useForm } from 'react-hook-form';
 import { useHeapState } from '@/state/heap/heap';
-import { useChannelFlag } from '@/hooks';
 import { isLinkCurio, isValidUrl } from '@/logic/utils';
 import useRequestState from '@/logic/useRequestState';
 import { JSONContent } from '@tiptap/core';
@@ -12,6 +11,7 @@ import { inlinesToJSON, JSONToInlines } from '@/logic/tiptap';
 import { ChatBlock } from '@/types/chat';
 import { Inline } from '@/types/content';
 import ConfirmationModal from '@/components/ConfirmationModal';
+import { useChannelFlag } from '@/logic/channel';
 import useCurioFromParams from './useCurioFromParams';
 import HeapTextInput from './HeapTextInput';
