+++ ui/src/logic/useAllBriefs.ts
@@ -2,15 +2,19 @@ import { useBriefs } from '@/state/chat';
 import { useBriefs as useHeapBriefs } from '@/state/heap/heap';
 import { useDiaryBriefs } from '@/state/diary';
 import _ from 'lodash';
+import { useMemo } from 'react';
 
 export default function useAllBriefs() {
   const chBriefs = useBriefs();
   const heBriefs = useHeapBriefs();
   const diBriefs = useDiaryBriefs();
 
-  return {
-    ..._.mapKeys(chBriefs, (v, k) => `chat/${k}`),
-    ..._.mapKeys(heBriefs, (v, k) => `heap/${k}`),
-    ..._.mapKeys(diBriefs, (v, k) => `diary/${k}`),
-  };
+  return useMemo(
+    () => ({
+      chat: chBriefs,
+      heap: heBriefs,
+      diary: diBriefs,
+    }),
+    [chBriefs, heBriefs, diBriefs]
+  );
 }
