--- ui/src/logic/useChannelSort.ts
@@ -1,67 +1,0 @@
-import { get } from 'lodash';
-import { useGroup, useRouteGroup } from '@/state/groups';
-import { GroupChannel, Channels, Zone } from '@/types/groups';
-import useSidebarSort, {
-  ALPHABETICAL,
-  DEFAULT,
-  RECENT,
-  sortAlphabetical,
-  Sorter,
-  useRecentSort,
-} from './useSidebarSort';
-
-const UNZONED = 'default';
-
-export default function useChannelSort() {
-  const groupFlag = useRouteGroup();
-  const group = useGroup(groupFlag);
-  const sortRecent = useRecentSort();
-
-  const sortDefault = (a: Zone, b: Zone) => {
-    if (!group) {
-      return 0;
-    }
-    const aIdx =
-      a in group['zone-ord']
-        ? group['zone-ord'].findIndex((e) => e === a)
-        : Number.POSITIVE_INFINITY;
-    const bIdx =
-      b in group['zone-ord']
-        ? group['zone-ord'].findIndex((e) => e === b)
-        : Number.POSITIVE_INFINITY;
-    return aIdx - bIdx;
-  };
-
-  const sortOptions: Record<string, Sorter> = {
-    [ALPHABETICAL]: sortAlphabetical,
-    [DEFAULT]: sortDefault,
-    [RECENT]: sortRecent,
-  };
-
-  const { sortFn, setSortFn, sortRecordsBy } = useSidebarSort({
-    sortOptions,
-    flag: groupFlag === '' ? '~' : groupFlag,
-  });
-
-  function sortChannels(channels: Channels) {
-    const accessors: Record<string, (k: string, v: GroupChannel) => string> = {
-      [ALPHABETICAL]: (_flag: string, channel: GroupChannel) =>
-        get(channel, 'meta.title'),
-      [DEFAULT]: (_flag: string, channel: GroupChannel) =>
-        channel.zone || UNZONED,
-      [RECENT]: (flag: string, _channel: GroupChannel) => flag,
-    };
-
-    return sortRecordsBy(channels, accessors[sortFn], sortFn === RECENT);
-  }
-
-  return {
-    setSortFn,
-    sortFn,
-    sortOptions: {
-      ...sortOptions,
-      [DEFAULT]: sortDefault,
-    },
-    sortChannels,
-  };
-}
