+++ ui/src/logic/useDismissChannelNotifications.ts
@@ -1,4 +1,4 @@
-import useIsChannelUnread from '@/logic/useIsChannelUnread';
+import { useIsChannelUnread } from '@/logic/channel';
 import { useNotifications } from '@/notifications/useNotifications';
 import { useRouteGroup } from '@/state/groups';
 import { useSawRopeMutation } from '@/state/hark';
