+++ ui/src/logic/useFilteredSections.ts
@@ -1,13 +1,12 @@
 import { useGroup, useVessel } from '@/state/groups';
-import useAllBriefs from './useAllBriefs';
-import useChannelSections from './useChannelSections';
-import { canReadChannel, isChannelJoined } from './utils';
+import { canReadChannel } from './utils';
+import { useChannelSections, useCheckChannelJoined } from './channel';
 
 const useFilteredSections = (flag: string, filterJoined?: boolean) => {
-  const briefs = useAllBriefs();
   const { sections, sectionedChannels } = useChannelSections(flag);
   const group = useGroup(flag);
   const vessel = useVessel(flag, window.our);
+  const isChannelJoined = useCheckChannelJoined();
 
   const filteredSections = sections.filter(
     (s) =>
@@ -17,7 +16,7 @@ const useFilteredSections = (flag: string, filterJoined?: boolean) => {
           return (
             sectionedChan[1] &&
             canReadChannel(sectionedChan[1], vessel, group?.bloc) &&
-            isChannelJoined(sectionedChan[0], briefs)
+            isChannelJoined(sectionedChan[0])
           );
         }
         return (
