--- ui/src/logic/useIsChannelHost.ts
@@ -1,3 +1,0 @@
-const useIsChannelHost = (flag: string) => window.our === flag?.split('/')[0];
-
-export default useIsChannelHost;
